refactor: chat leave dto removed
