+++ backend/src/socket/chat/service/chat-user.service.ts
@@ -41,38 +41,6 @@ export class ChatUserService {
     });
   }
 
-  public leaveChatRoom(chatRoomId: string, socketId: string): void {
-    const chatRoom = this.chatRoomRepository.getChatRoom(chatRoomId);
-
-    if (!chatRoom) {
-      // throw new NotAcceptableException(`Chat room ${chatRoomId} not found`);
-      return;
-    }
-
-    if (!chatRoom.sockets.has(socketId)) {
-      // throw new NotAcceptableException(`Socket ${socketId} is not in chat room ${chatRoomId}`);
-      return;
-    }
-
-    if (chatRoom.sockets.size === 1) {
-      this.chatRoomRepository.removeChatRoom(chatRoomId);
-      return;
-    }
-
-    if (chatRoom.sockets.get(socketId).role === 'host') {
-      const nextHost = Array.from(chatRoom.sockets.values()).find(chatUser => chatUser.role === 'operator');
-
-      if (nextHost) {
-        nextHost.role = 'host';
-      } else {
-        const randomHost = Array.from(chatRoom.sockets.values())[0];
-        randomHost.role = 'host';
-      }
-    }
-
-    this.chatRoomRepository.removeSocketFromChatRoom(chatRoomId, socketId);
-  }
-
   public leaveAllChatRooms(socketId: string): void {
     this.chatRoomRepository.removeSocketFromAllChatRoom(socketId);
   }
